refactor: common `splitWebsiteField`
